+++ shuup/admin/modules/orders/sections.py
@@ -79,23 +79,31 @@ class ShipmentSection(Section):
         suppliers = Supplier.objects.filter(order_lines__order=order).distinct()
         create_permission = "order.create-shipment"
         delete_permission = "order.delete-shipment"
-        missing_permissions = get_missing_permissions(request.user, [create_permission, delete_permission])
+        set_sent_permission = "order.set-shipment-sent"
+        missing_permissions = get_missing_permissions(
+            request.user,
+            [create_permission, delete_permission, set_sent_permission]
+        )
         create_urls = {}
+        delete_urls = {}
+        set_sent_urls = {}
+
         if create_permission not in missing_permissions:
             for supplier in suppliers:
                 create_urls[supplier.pk] = reverse(
                     "shuup_admin:order.create-shipment", kwargs={"pk": order.pk, "supplier_pk": supplier.pk})
 
-        delete_urls = {}
-        if delete_permission not in missing_permissions:
-            for shipment_id in order.shipments.all_except_deleted().values_list("id", flat=True):
-                delete_urls[shipment_id] = reverse(
-                    "shuup_admin:order.delete-shipment", kwargs={"pk": shipment_id})
+        for shipment_id in order.shipments.all_except_deleted().values_list("id", flat=True):
+            if delete_permission not in missing_permissions:
+                delete_urls[shipment_id] = reverse("shuup_admin:order.delete-shipment", kwargs={"pk": shipment_id})
+            if set_sent_permission not in missing_permissions:
+                set_sent_urls[shipment_id] = reverse("shuup_admin:order.set-shipment-sent", kwargs={"pk": shipment_id})
 
         return {
             "suppliers": suppliers,
             "create_urls": create_urls,
-            "delete_urls": delete_urls
+            "delete_urls": delete_urls,
+            "set_sent_urls": set_sent_urls,
         }