+++ shuup/core/signals.py
@@ -12,6 +12,7 @@ get_visibility_errors = Signal(providing_args=["shop_product", "customer"], use_
 get_orderability_errors = Signal(providing_args=["shop_product", "customer", "supplier", "quantity"], use_caching=True)
 shipment_created = Signal(providing_args=["order", "shipment"], use_caching=True)
 shipment_created_and_processed = Signal(providing_args=["order", "shipment"], use_caching=True)
+shipment_sent = Signal(providing_args=["order", "shipment"], use_caching=True)
 refund_created = Signal(providing_args=["order", "refund_lines"], use_caching=True)
 category_deleted = Signal(providing_args=["category"], use_caching=True)
 shipment_deleted = Signal(providing_args=["shipment"], use_caching=True)